Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main(){
-    float radius,area;
-    scanf("%f", &radius);
+    int radius,area;
+    scanf("%d", &radius);
     Area=3.14(radius*radius);
-    printf("Area: %.2f", Area);
+    printf("Area: %d", Area);
     return 0;
 }
